feat(scripts): set up dev and start scripts, order dependencies, change README
