+++ .gitignore
@@ -9,6 +9,7 @@
 /coverage
 /artifacts
 /cache
+.temp
 
 # production
 /build
