+++ package.json
@@ -3,10 +3,10 @@
   "version": "1.0.10",
   "description": "dxvote application of DXdao.",
   "scripts": {
+    "dev": "./scripts/dev.sh",
     "start": "./scripts/start.sh",
     "build": "./scripts/build.sh",
     "build-analyze": "yarn cra-bundle-analyzer",
-    "dev": "dotenv -e .env.development ./scripts/dev.sh",
     "lint": "eslint src/",
     "format": "prettier --write \"src/**/*.{js,jsx,ts,tsx,json,md}\" \"scripts/*.{js,jsx,ts,tsx,json,md}\"",
     "format-check": "prettier --check \"src/**/*.{js,jsx,ts,tsx,json,md}\" \"scripts/*.{js,jsx,ts,tsx,json,md}\"",
@@ -31,6 +31,10 @@
   },
   "dependencies": {
     "@0xsequence/multicall": "^0.35.5",
+    "@nomiclabs/hardhat-truffle5": "^2.0.5",
+    "@nomiclabs/hardhat-web3": "^2.0.0",
+    "@testing-library/jest-dom": "^5.16.2",
+    "@testing-library/react": "^12.1.4",
     "@tiptap/extension-focus": "^2.0.0-beta.39",
     "@tiptap/extension-highlight": "^2.0.0-beta.32",
     "@tiptap/extension-placeholder": "^2.0.0-beta.48",
@@ -49,11 +53,14 @@
     "bignumber.js": "^9.0.1",
     "content-hash": "^2.5.2",
     "copy-to-clipboard": "^3.3.1",
-    "dxdao-contracts": "https://github.com/DXgovernance/dxdao-contracts.git#b7eb732c0c3c81ea53ed928a3077d71781941b50",
+    "dxdao-contracts": "https://github.com/AugustoL/dxdao-contracts.git#37a8dbec9e7e744a712979b4a531f11b4e93eb82",
+    "eslint-plugin-cypress": "^2.12.1",
     "ether-swr": "^2.0.1",
     "ethers": "^5.5.2",
     "file-saver": "^2.0.5",
     "git-revision-webpack-plugin": "^3.0.6",
+    "hardhat": "^2.9.1",
+    "hardhat-dependency-compiler": "^1.1.2",
     "ipfs": "^0.58.0",
     "ipfs-only-hash": "^4.0.0",
     "json-keys-sort": "^2.1.0",
@@ -89,35 +96,27 @@
     "sanitize-html": "^2.6.1",
     "styled-components": "^5.3.0",
     "swr": "^1.1.2",
+    "ts-generator": "^0.1.1",
     "ts-node": "^9.1.1",
     "turndown": "^7.1.1",
+    "typechain-target-ethers-v5": "^5.0.1",
     "typescript": "^4.5.5",
     "webpack": "4.44.2",
     "webpack-manifest-plugin": "^2.2.0",
     "workbox-webpack-plugin": "^5.1.4"
   },
   "devDependencies": {
-    "@nomiclabs/hardhat-truffle5": "^2.0.0",
-    "@nomiclabs/hardhat-web3": "^2.0.0",
-    "@testing-library/cypress": "^8.0.2",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^12.1.2",
-    "@typechain/ethers-v5": "^8.0.5",
+    "@typechain/ethers-v5": "^9.0.0",
     "@types/jest": "^27.4.0",
     "@types/jsonexport": "^3.0.2",
     "@types/react-router-dom": "^5.1.2",
     "cra-bundle-analyzer": "^0.1.0",
     "cypress": "^9.2.0",
-    "dotenv": "^8.1.0",
-    "dotenv-cli": "^4.0.0",
-    "eslint-plugin-cypress": "^2.12.1",
-    "hardhat": "^2.2.0",
-    "hardhat-dependency-compiler": "^1.1.1",
+    "husky": "^6.0.0",
     "prettier": "2.5.0",
     "start-server-and-test": "^1.14.0",
-    "typechain": "^6.0.5",
-    "web3": "^1.3.5",
-    "husky": "^6.0.0"
+    "typechain": "^7.0.1",
+    "web3": "^1.3.5"
   },
   "browserslist": {
     "production": [
