+++ scripts/deployDevConfig.json
@@ -1,0 +1,181 @@
+{
+  "permissionRegistryDelay": 3600,
+  "contributionReward": {
+    "queuedVoteRequiredPercentage": 50,
+    "queuedVotePeriodLimit": 600,
+    "boostedVotePeriodLimit": 300,
+    "preBoostedVotePeriodLimit": 150,
+    "thresholdConst": 2000,
+    "quietEndingPeriod": 150,
+    "proposingRepReward": 10,
+    "votersReputationLossRatio": 100,
+    "minimumDaoBounty": 100000000000000000000,
+    "daoBountyConst": 100
+  },
+  "walletSchemes": [
+    {
+      "name": "RegistrarWalletScheme",
+      "doAvatarGenericCalls": true,
+      "maxSecondsForExecution": 86400,
+      "maxRepPercentageChange": 0,
+      "controllerPermissions": {
+        "canGenericCall": true,
+        "canUpgrade": true,
+        "canRegisterSchemes": true
+      },
+      "permissions": [],
+      "queuedVoteRequiredPercentage": 75,
+      "boostedVoteRequiredPercentage": 500,
+      "queuedVotePeriodLimit": 600,
+      "boostedVotePeriodLimit": 300,
+      "preBoostedVotePeriodLimit": 150,
+      "thresholdConst": 2000,
+      "quietEndingPeriod": 150,
+      "proposingRepReward": 0,
+      "votersReputationLossRatio": 100,
+      "minimumDaoBounty": 100000000000000000000,
+      "daoBountyConst": 100
+    },
+    {
+      "name": "MasterWalletScheme",
+      "doAvatarGenericCalls": true,
+      "maxSecondsForExecution": 86400,
+      "maxRepPercentageChange": 40,
+      "controllerPermissions": {
+        "canGenericCall": true,
+        "canUpgrade": false,
+        "canChangeConstraints": false,
+        "canRegisterSchemes": false
+      },
+      "permissions": [
+        {
+          "asset": "0x0000000000000000000000000000000000000000",
+          "to": "DXDVotingMachine",
+          "functionSignature": "0xaaaaaaaa",
+          "value": "0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff",
+          "allowed": true
+        },
+        {
+          "asset": "0x0000000000000000000000000000000000000000",
+          "to": "RegistrarWalletScheme",
+          "functionSignature": "0xaaaaaaaa",
+          "value": "0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff",
+          "allowed": true
+        },
+        {
+          "asset": "0x0000000000000000000000000000000000000000",
+          "to": "0xaAaAaAaaAaAaAaaAaAAAAAAAAaaaAaAaAaaAaaAa",
+          "functionSignature": "0xaaaaaaaa",
+          "value": "0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff",
+          "allowed": true
+        }
+      ],
+      "queuedVoteRequiredPercentage": 50,
+      "boostedVoteRequiredPercentage": 200,
+      "queuedVotePeriodLimit": 300,
+      "boostedVotePeriodLimit": 90,
+      "preBoostedVotePeriodLimit": 30,
+      "thresholdConst": 1500,
+      "quietEndingPeriod": 60,
+      "proposingRepReward": 0,
+      "votersReputationLossRatio": 5,
+      "minimumDaoBounty": 100000000000000000,
+      "daoBountyConst": 10
+    },
+    {
+      "name": "QuickWalletScheme",
+      "doAvatarGenericCalls": false,
+      "maxSecondsForExecution": 86400,
+      "maxRepPercentageChange": 1,
+      "controllerPermissions": {
+        "canGenericCall": false,
+        "canUpgrade": false,
+        "canChangeConstraints": false,
+        "canRegisterSchemes": false
+      },
+      "permissions": [
+        {
+          "asset": "0x0000000000000000000000000000000000000000",
+          "to": "0xaAaAaAaaAaAaAaaAaAAAAAAAAaaaAaAaAaaAaaAa",
+          "functionSignature": "0xaaaaaaaa",
+          "value": "0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff",
+          "allowed": true
+        },
+        {
+          "asset": "DXD",
+          "to": "0xaAaAaAaaAaAaAaaAaAAAAAAAAaaaAaAaAaaAaaAa",
+          "functionSignature": "0xaaaaaaaa",
+          "value": "0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff",
+          "allowed": true
+        }
+      ],
+      "queuedVoteRequiredPercentage": 50,
+      "boostedVoteRequiredPercentage": 1000,
+      "queuedVotePeriodLimit": 300,
+      "boostedVotePeriodLimit": 60,
+      "preBoostedVotePeriodLimit": 30,
+      "thresholdConst": 1300,
+      "quietEndingPeriod": 30,
+      "proposingRepReward": 0,
+      "votersReputationLossRatio": 10,
+      "minimumDaoBounty": 100000000000000000,
+      "daoBountyConst": 10
+    }
+  ],
+  "reputation": [
+    {
+      "address": "0x79706c8e413cdaee9e63f282507287b9ea9c0928",
+      "amount": 51000000000
+    },
+    {
+      "address": "0xc73480525e9d1198d448ece4a01daea851f72a9d",
+      "amount": 48000000000
+    },
+    {
+      "address": "0x3f943f38b2fbe1ee5daf0516cecfe4e0f8734351",
+      "amount": 1000000000
+    }
+  ],
+
+  "actions": [
+    {
+      "type": "proposal",
+      "time": 1,
+      "from": "0x3f943f38b2fbe1ee5daf0516cecfe4e0f8734351",
+      "data": {
+        "to": ["0x3f943f38b2fbe1ee5daf0516cecfe4e0f8734351"],
+        "callData": ["0x00"],
+        "value": ["1000000000000000000000000000000000000000000"],
+        "title": "Proposal Test #0",
+        "description": "Send 1000 (in WEI) ETH to 0x3f943f38b2fbe1ee5daf0516cecfe4e0f8734351",
+        "tags": ["dxvote"],
+        "scheme": "MasterWalletScheme"
+      }
+    },{
+      "type": "stake",
+      "time": 2,
+      "from": "0x3f943f38b2fbe1ee5daf0516cecfe4e0f8734351",
+      "data": {
+        "proposal": "0",
+        "decision": "1",
+        "amount": "200000000000000000"
+      }
+    },{
+      "type": "vote",
+      "time": 33,
+      "from": "0xc73480525e9d1198d448ece4a01daea851f72a9d",
+      "data": {
+        "proposal": "0",
+        "decision": "1",
+        "amount": "0"
+      }
+    },{
+      "type": "execute",
+      "time": 101,
+      "from": "0x3f943f38b2fbe1ee5daf0516cecfe4e0f8734351",
+      "data": {
+        "proposal": "0"
+      }
+    }
+  ]
+}
